Data.Array.Accelerate.IO.Data.Array.IArray

Description

Convert between immutable IArrays and Accelerate Arrays.

Documentation

type family IxShapeRepr e where ...

Equations

IxShapeRepr () = ()

IxShapeRepr Int = ((), Int)

IxShapeRepr (t, h) = (IxShapeRepr t, h)

fromIArray :: (IxShapeRepr (EltRepr ix) ~ EltRepr sh, IArray a e, Ix ix, Shape sh, Elt ix, Elt e) => a ix e -> Array sh e

O(n). Convert an IArray to an Accelerate Array.

The index type ix of the IArray corresponds to the shape sh of the Accelerate Array in the following way:

DIM0 ~ ()
DIM1 ~ Int
DIM2 ~ (Int,Int)
DIM3 ~ (Int,Int,Int)

...and so forth.

toIArray

Arguments

:: (IxShapeRepr (EltRepr ix) ~ EltRepr sh, IArray a e, Ix ix, Shape sh, Elt ix)

=> Maybe ix
if Just this as the index lower bound, otherwise the array is indexed from zero

-> Array sh e

-> a ix e

O(n). Convert an Accelerate Array to an IArray.

See fromIArray for a discussion on the expected shape types.