Data.Array.Accelerate.IO.Data.Vector.Unboxed

Description

Efficient conversion between Unboxed vectors and Accelerate Arrays.

Documentation

class (Unbox e, Elt e) => Unbox e where

Minimal complete definition

arrayDataOfUnboxed, unboxedOfArrayData

Methods

arrayDataOfUnboxed :: Vector e -> ArrayData (EltRepr e)

unboxedOfArrayData :: Int -> ArrayData (EltRepr e) -> Vector e

toUnboxed :: (Shape sh, Unbox e) => Array sh e -> Vector e

O(1) (typically). Convert an Accelerate array into an Unboxed vector.

If the array data was allocated by Accelerate, this can typically be done without copying. The resulting vector will be pinned.

since 1.1.0.0

fromUnboxed :: Unbox e => Vector e -> Array DIM1 e

O(n) (typically). Convert an Unboxed vector into an Accelerate array Accelerate array.

If the underlying vectors are pinned then this can be done without copying.

See also: https://ghc.haskell.org/trac/ghc/ticket/5556

since 1.1.0.0