Data.Array.Accelerate.IO.Foreign.ForeignPtr

Description

Documentation

type family ForeignPtrs e

A family of types which represent a collection of ForeignPtrs. The structure of the collection depends on the element type e.

fromForeignPtrs :: (Shape sh, Elt e) => sh -> ForeignPtrs (EltRepr e) -> Array sh e

O(1). Treat the set of ForeignPtrs as an Accelerate array. The type of elements e in the output Accelerate array determines the structure of the collection.

Data is considered to be in row-major order. You must ensure that each of the input pointers contains the right number of elements.

The data may not be modified through the ForeignPtrs afterwards.

You should make sure that the data is suitably aligned.

since 1.1.0.0

toForeignPtrs :: (Shape sh, Elt e) => Array sh e -> ForeignPtrs (EltRepr e)

O(1). Yield the ForeignPtrs underlying the given Accelerate Array. The element type e will determine the structure of the output collection.

Data is considered to be in row-major order.

since 1.1.0.0